% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateAnimalModelInput.R
\name{genAnimalModelInput}
\alias{genAnimalModelInput}
\title{Generate input for BSFG model in Matlab and Julia}
\usage{
genAnimalModelInput(formula, data, A, IDcol = "ID", out_folder = NULL)
}
\arguments{
\item{formula}{formula specifing fixed effects and response variables}

\item{data}{data.frame with data, default is ratonesdf}

\item{A}{relatinship matrix from pedigree}

\item{out_folder}{folder to create the Matlab input files}
}
\value{
list with animal model input objects
}
\description{
Generate input for BSFG model in Matlab and Julia
}
\examples{
library(ratones)
data(ratonesdf)
data(ratones_ped)
formula = paste0("cbind(",
          paste(names(dplyr::select(ratonesdf, IS_PM:BA_OPI)),
                collapse = ", "), ") ~ SEX + AGE")
animal_model_data = generateAnimalModelInput(formula, ratonesdf,
                                             ratones_ped$A, out_foldel = "./")
}
